+++ CHANGELOG.md
@@ -10,12 +10,21 @@ and this project adheres to
 
 ### Added
 
+- Added ingestion of Azure Management Groups when the
+  `configureSubscriptionInstances` configuration field is `true`. This ingestion
+  requires users to assign the `Management Group Reader`
+  [role](https://docs.microsoft.com/en-us/azure/role-based-access-control/built-in-roles#management-group-reader)
+  to the service principal used in the given integration configuration to the
+  **Tenant Root Group**
+  [management group](https://docs.microsoft.com/en-us/azure/governance/management-groups/overview#root-management-group-for-each-directory).
+
 - Added support for ingesting the following **new** resources:
 
-  | Service       | Resource / Entity    |
-  | ------------- | -------------------- |
-  | Gallery       | `azure_gallery`      |
-  | Gallery Image | `azure_shared_image` |
+  | Service           | Resource / Entity        |
+  | ----------------- | ------------------------ |
+  | Gallery           | `azure_gallery`          |
+  | Gallery Image     | `azure_shared_image`     |
+  | Management Groups | `azure_management_group` |
 
 - Added support for ingesting the following **new** relationships:
 
@@ -25,6 +34,8 @@ and this project adheres to
   | `azure_gallery`        | `CONTAINS` | `azure_shared_image` |
   | `azure_vm`             | `HAS`      | `azure_shared_image` |
   | `azure_vm`             | `HAS`      | `azure_image`        |
+  | `azure_management_group` | `CONTAINS` | `azure_management_group` |
+  | `azure_account`          | `HAS`      | `azure_management_group` |
 
 - New properties added to resources: